Sync minor changes
