--- guild permission levels
@@ -1,10 +1,0 @@
-
-OWNER
-
-ADMIN
-
-MODERATOR
-
-MEMBER
-
-BLACKLISTED
