+++ index.js
@@ -5,9 +5,6 @@ const { Client, Collection, GatewayIntentBits, Partials } = require('discord.js'
 const ModuleManager = require('./structures/ModuleManager.js');
 const Database = require('./structures/Database.js');
 const ConfigurationManager = require('./structures/ConfigurationManager.js');
-const SetOwners = require('./modules/SetOwners.js')
-const Module = require('./structures/Module.js')
-const fs = require('fs');
 BigInt.prototype.toJSON = function() { return this.toString() } // MDN https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt#use_within_json
 
 // Discord
